Fixed init script
