--- cam_file_server/Build.sh
@@ -1,1 +1,0 @@
-go build cam_file_server.go
