--- cam_file_server/go.mod
@@ -1,3 +1,0 @@
-module receive_file
-
-go 1.20
