+++ cam_server/Build.sh
@@ -1,0 +1,1 @@
+go build cam_server.go
