+++ cam_server/go.mod
@@ -1,0 +1,3 @@
+module receive_file
+
+go 1.20
